gpu/requirements.txt
fastapi
uvicorn
pydantic
fastembed-gpu